+++ Changelog.md
@@ -1,6 +1,17 @@
 # Changelog
 
 
+### 1.1.0 "Accessorized" (3/-/23)
+* Added Hunting Shotgun.
+* Added Extended Barrel, Flash Hider, Muzzle Break, and Heavy Muzzle Break.
+* Added Advanced Grip.
+* Added Carbine Stock.
+* Buffed the stats of several guns to play better with unmodified CGM.
+* Tweaked several gun models.
+
+NineZero's Gun Expansion now requires CGM 1.3.4; if you are running an older CGM version, please update to 1.3.4.
+
+
 ### 1.0.1 (2/1/23)
 * Fixed invalid crafting recipes for stocks.